Fix inability to alter microcode pkg once set
This loops on cpuAnswer rather than cpuPackage, the latter of which will
always skip the loop after the first run once a valid answer's selected.
We also replace the if/else with a case to clean things up a little.

To make sure the loop always resets when an invalid or empty answer is
given, we reset the value for cpuAnswer in the case's default section.