Rewrite target disk setting to loop and use option
We can't use `$#` and `$1` because the function receives all args, as
well as ones already parsed. Instead, this rewrites set_target_disk so
that it prompts recursively until a valid block device is given.

With a while loop enforcing a block device from the input, this will now
continue to prompt when bad input is given, but it also has the drawback
in that "/dev/sda1" passes as valid, but we'd actually want "/dev/sda".

To retain setting the device in the command, $1 becomes $TARGET_DEVICE,
set by the newly added `-d` or `--device` option. In other words:
`dashit /dev/nvme0n1` becomes `dashit -d /dev/nvme0n1`.

This resolves errors when setting a device with `--dry-run` for example.