Add account creation
Creates the user account and adds it to sudoers, after also installing
sudo with man-db and the man pages added to the list too.

Once the user is created, we expire the password to force it being set
at the next login, which will presumably be straight after the reboot.