Fix refind config
Without setting the allowed names explicitly, refind doesn't know how to
match Arch's kernel names to their initramfs images in order to perform
its replacements on "%v" in the initrd path. As a result, refind will be
unable to boot the system because it's looking for "initramfs-.img" when
it should actually be referencing "initramfs-linux.img" instead.

The extra_kernel_version_strings option is disabled in default refind
config, so this enables it and adds other valid arch kernel names.