Install yay after preparing pacman
Adds a function that clones the binary verison of yay into /opt in the
chroot, and sets nobody as owner so that it can be built without root,
before eventually being installed with a pacman -U $pkgFile.

In order to clone and run makepkg etc, this also reshuffles the default
packages list slightly, moving dhcpcd, refind and reflector to a new
line along with the needed base-devel and git packages.