Auto-retry if pacstrap fails
Occasionally, pacstrap fails with an error that some packages failed
verification and either the package itself, or key used to verify it,
are invalid or potentially corrupted.

If I then run install a second time (but manual mode to avoid
reformatting) pacstrap seems to simply pick up where it left off,
working fine to install the rest of the packages, so this does that.

Also cleans up a couple other bits that we don't seem to need.