Add GH release workflow, ignore dashit and *.md
Adds the "compiled" `dashit` script to gitignore along with all markdown
files except the readme so it can stop complaining about my local notes.

The release workflow runs when a `v*.*.*` tag is pushed. When triggered,
it will run `make` before creating a release with the resulting `dashit`
script uploaded as a compiled asset.

Only problem with this method is it always defaults to the tag name for
the release's name, but so do all the alternatives. As a result, GitHub
is prevented from setting it automatically from the tag message.

The release name can be removed manually in GH though, so no big deal.