Pacman: Only install required packages
The arch ISO already has everything we need from arch-install-scripts,
so this adds the `--needed` flag so we can still get it on arch-based
distros or an existing Arch system when installing to a second drive.

As the Arch ISO uses a limited partition size by default, full system
upgrades will often fail, particularly if there's a kernel update. To
avoid issues, this commit also skips `pacman -Syyu` on ArchISO hosts.