Root: Copy host's password using systemd-firstboot
Replaces `passwd` with a call to `systemd-firstboot` with params set to
copy the root password from the host machine's shadow file, and adds an
"empty password" check to both manual and auto install modes.

If the check finds root has no password on the host, it'll ask the user
if they want to set one before proceeding with install. If passwords
don't match, it'll continue prompting until they do, or user says no.