Pacstrap: Add initramfs package selection menu
The linux package requires `initramfs` rather than one specific package,
so Pacman will by default install all of them. This adds a menu so that
a specific one can be chosen if so desired. Defaults to mkinitcpio, but
only once the selection menu has been entered. Default is all otherwise.