First Boot: Support custom user and root scripts
Adds the ability to have DASHit automatically run two scripts on first
boot - one as root and the other as the newly created system user.

Scripts should be placed in a "dashit" directory in the root of your
dotfiles and named `firstboot.(user|root).bash` to be detected. In
addition, dotfiles must be specified using `DASHIT_DOTFILES_REPO`.