First Boot: Create an .xinitrc if it doesn't exist
Adds picom and numlockx to the list of packages for i3, and adds them
along with redshift and the exec line to a basic .xinitrc if there isn't
already one in the user's homedir after running the custom user script.

This also adds a prompt to return to tty1 after successful install. It
doesn't work settin `ExecStartPost=chvt 1` because it needs root, and
the Pre/Post commands are run as the same user as the ExecStart script.