Disk: Partition using sgdisk for GPT automounting
Parted doesn't allow for setting fstype by GUID or typecode, only a set
of values that it then interprets. This commit switches to sgpart using
a filesystem typecode of 8304 which is a "Linux x86-64 root" and should
support automounting via the Discoverable Partitions Specification.

If you're reading this in git blame:
* x86: 8303
* ARM: 8307
* A64: 8305