Disk: Remove root partition and ESP from fstab
This removes comments, blank lines and the ESP/root partitions from the
fstab, and moves the root partition's mount options into `rootflags`.

This wouldn't be bootable without telling Refind that it should write
systemd's EFI "variables" (really it's only one) so this commit also
updates the refind.conf appropriately. Logic is identical as the other
tweak to the refind settings, so we move it to its own function here.